Changed and updated some styles.
Also added a secondary search in combination with the original search function.

Now the page will search each text of each element for the regex, but also each elements innerHTML as a string for the match.
There were some situations in which matches were not found because there was html elements inside the string of text, but it was presented on the page as regular text, and should be matched.